Scale new nodes correctly so that LaTeX-font sizes match
To ensure that the selected LaTeX font size in a new node matches the
font size in Inkscape the scaling of the node has to be properly chosen.

From Inkscape 0.92 on the user unit of the document is mm per default. This
means that the svg coordinates in the new node coming from pstoedit are
interpreted as mm, too. However, their are meant as pt in pstoedit.
Hence, when the user selects a scale factor of 1.0 a font size of eg.
12pt will be displayed 3 times larger as expected (1mm = 2.83pt).

To get rid of this problem, we have to scale the new node appropriately
which can be done by calculating the required conversion factor from the
method unittouu. When the user selects 1.0 as scale factor it is
internally treated as 0.3... . Different user units (e.g. inch) are
handled automatically by unittouu.

In Inkscape <= 0.91 the default unit was pt so when the user selected
1.0 as scale factor the font size in the new node was correct.