Fixed correct declaration of abstract static methods
In Python 2.7 (as currently used by Inkscape) abstract static methods
must be declared as @abc.abstractmethod in the abstract base class and
as @staticmethod in the derived classes. The syntax

@staticmethod
@abs.abstractmethod
def xyz()

known from Python >= 3.3 does not work properly. (It does not cause any
runtime errors but if you forget to implement a function in a derived
class no exception is thrown during object instantiation).