Renamed all instances of SVGElement from node to svg_ele
To ensure that one gets not confused if an object is an instance of
SvgElement or an instance of xmltree all objects which are instances of
SvgElement and its subclasses contain "svg_ele" in their name.

Mainly this means: old_node -> old_svg_ele, new_node -> new_svg_ele

The names of instances of xmltree keep their original names (old_node,
new_node and of course _node in SvgElement).