Processing of xelatex and lualatex commands enabled
The TeX command selected in the GUI is now passed to the
PdfConverterBase class and executed in the tex_to_pdf method.

When re-editing a node the command used for creating the node in the
past is passed to the GUI.

Be aware that LuaLaTex on Windows does not accept backslashes in the
filename. Hence, we have to account for that in the tex_to_pdf method