SvgElement: Added methods for TeX color detection and color style import
Ensures the following: If colorized TeX code is used the node is
inserted in the document with the TeX colors, if not, the group color defined
in Inkscape is used (incl. opacity). Specific colorization of single characters
done in Inkscape is intentionally lost.

New method is_colorized() returns True if at least one element with a
non-black color is found in the node. Specific implementations for
pdf2svg and pstoedit (pdf2svg uses css styles, pstoedit uses
attributes).

Can be used to check if colors have been explicitely set in the TeX
code immediately after creation of the SVG node.

New method import_group_color_style which imports the color style information
from an existing SVGElement and applies it to all items of the managed
node.