More reliable deletion of temp directory
Directory is also deleted if it has subdirectories or contains files
which do not begin with tmp. It seems that some tex distributions
produce additional output compared to what has been observed so far.
Hence, this patch forces deletion of the temp directory even if it is
not empty.

See issue #34