Added possibility to check multiple executable names
Under Windows several executable names exist for the same program, e.g.
gs.exe, gswin32c.exe or gswin64c.exe for ghostscript. This commit
accounts for this problem. For this purpose the executable names can be
either given as a single string or as a list of strings in the
properties of the Default class. The method find_executable of TexTextRequirementsChecker accepts a string or a list of strings as argument and checks all given executable names instead of only one as in the previous version. Same holds for _find_executable_in_path.