Enabled colored output on Windows 10 since anniversary update
The python distribution shipped with Inkscape on Windows does not offer
the colorama package which would usually be used for colored output on
any Windows system.

Fortunately, Windows 10 offers the possibility to use ANSI sequences in
the command window since the anniversary update (build 14393). However,
it must be explicitely enabled by default via the Windows API. See e.g.
https://wpdev.uservoice.com/forums/266908-command-prompt-console-bash-on-ubuntu-on-windo/suggestions/15617610--re-enable-enable-virtual-terminal-processing-by

Hence, we check the Windows version and try to enable the colorization
if we run on the correct Windows version.