Removed -psarg -r9600x9600 option in pstoedit call
pstoedit 3.74 fixed an incompatibility with ghostscript 9.27. However,
the produced svg code is not rendered to what is expected when the
option -psarg -r9600x9600 is passed to the pstoedit call. Hence, it is
now removed. The result looks good, also with versions of pstoedit <
3.74 and gs < 9.27.

See also https://sourceforge.net/p/pstoedit/bugs/25/ and issue #126

Resolves #126