Adjusted tests to account for removed resolution flag in pstoedit conv.
In order to use pstoedit 3.74 together with gs 9.27 one requires to
remove the -psarg -r9600x9600 argument (see commit e2f4f9c6bd38bf2f6c0d663df2a1342cf8d53ae5
 and issue #126)

This results in very slight changes in the output so the fuzz parameter had to
be adapted in the tests. Manual checks show that the differences
allowed now in the tests are acceptable.