requirements_check: Adapted to work with alpha app image of Inkscape
For testing requirements check first checks for Inkscape alpha2 app
image, then for native inkscpae executable. Enables one to have stable
and alpha Inkscape in parallel.

Not that alpha2 image returns two lines when invoked with the --version
option, hence the corresponding code had to be modified in a more
general manner.