Fixed: Inkscape version never stored in TexText node
Since exception thrown when acessing nonexising self.svg member
is silently ignored, the error did not show up during testing.

Furthermore the correct tag to query for is inkscape:version, not
version (have we been in the inkscape namespace by default in old
TexText verions?)

Finally set_meta is called instead of set so that namespace is correctly
set.