Fixed: Wrong placement of new nodes
New nodes were not being placed in the center of the document since
the return values of the bounding box method were incorrectly expanded
(x, y, w, h) instead of (x, w, y, h) and document and node units were
mixed (mm and pt).

New nodes are now placed in the view center instead in the center of the
document.