Explicit setting of log dir for  setup
This is due the way setup.py is called from Windows:
During installation on windows setup.py is called from
setup_win.bat. In order to enable python.exe shipped
with Inkscape to detect all libraries it must be run
from within the directory python.exe resides in. This
directory is read-only for standard users. Hence, we
explicitely set the log dir to the directory in which
setup.py resides.

This can only be a workaround. Better approach is to
setup a correct environment from setup_win.bat