Windows installer: Better handling of MSI Inkscape installations
When Inkscape is installed on Windows using the MSI installer the
registry keys are not set as it is done when using the NSIS installer.
In fact, no valid registry keys are set from which anything about the
install location can be concluded. Hence a simple trial and error method
has been added to the batch file script to resolve this in the most
likely cases.